requirements.txt
azure-storage-blob
fastapi==0.135.1
fonttools==4.61.1
langchain==1.2.13
langchain_core==1.2.20
langgraph==1.1.3
mssql_python==1.4.0
openai==2.29.0
pandas==3.0.1
plotly==6.6.0
pydantic==2.12.5
Requests==2.32.5
typing_extensions==4.15.0
python-dotenv
azure-cognitiveservices-speech
uvicorn
langchain-openai
python-multipart
azure-ai-documentintelligence 